+++ ExpCurvFitting/ExpCurvFitting.Application/Excel/IExcelService.cs
@@ -1,0 +1,8 @@
+﻿
+namespace ExpCurvFitting.Application
+{
+    public interface IExcelService
+    {
+        Result LoadDateFromFile(Stream file);
+    }
+}
