+++ ExpCurvFitting/ExpCurvFitting.Application/ExpCurvFitting.Application.csproj
@@ -7,7 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-	  <PackageReference Include="EPPlus" Version="4.5.3.3" />
+	  <PackageReference Include="ExcelDataReader" Version="3.7.0" />
+	  <PackageReference Include="ExcelDataReader.DataSet" Version="3.7.0" />
   </ItemGroup>
 
   <ItemGroup>
