+++ ExpCurvFitting/ExpCurvFitting.Core/Models/ExpWithMixinModel.cs
@@ -2,8 +2,6 @@
 using ExpCurvFitting.Core.Optimization;
 using MathNet.Numerics.LinearAlgebra;
 using MathNet.Numerics.LinearAlgebra.Double;
-using System.Diagnostics;
-using System.Drawing;
 
 namespace ExpCurvFitting.Core.Models;
 public record ExpWithMixinModel
