+++ ExpCurvFitting/ExpCurvFitting.Web/Components/Pages/ExpWithMixin.razor
@@ -1,13 +1,14 @@
 ﻿@page "/expfitting"
 @rendermode InteractiveServer
 @implements IDisposable
-@using Microsoft.AspNetCore.Components.Forms
-@using ExpCurvFitting.Application
-@using ExpCurvFitting.Core.Models
-@using ExpCurvFitting.Core.RecognizingFunctions
-@using ExpCurvFitting.Web.Models
-@using Serilog;
-@inject ExcelService _excelService
+@    using Microsoft.AspNetCore.Components.Forms
+@    using ExpCurvFitting.Application
+@    using ExpCurvFitting.Core.Models
+@    using ExpCurvFitting.Core.RecognizingFunctions
+@    using ExpCurvFitting.Web.Models
+@    using Serilog;
+    using ExpCurvFitting.Application.Excel;
+@inject IExcelService _excelService
 
 
 <PageTitle>Exponents curve fitting</PageTitle>
@@ -201,7 +202,7 @@ else
 
     private CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
     private int numberOfExponents;
-    private ExcelService.Result inputData = null!;
+    private Result inputData = null!;
     private ExpWithMixinModel.Result modelResult = null!;
     private OptimizationOptions optimizationOptions = new();
     private bool IsLoading;
